configurationmvc file added and problem fix

+++ Smart_Contact_manager/src/main/java/com/smart/contact/dao/UserRepository.java
@@ -1,14 +1,11 @@
 package com.smart.contact.dao;
 
-import org.springframework.boot.autoconfigure.domain.EntityScan;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
-import org.springframework.stereotype.Component;
 import org.springframework.stereotype.Repository;
 
 import com.smart.contact.entities.UserEntity;
 
 @Repository 
-public interface UserRepository extends JpaRepository<UserEntity, Long>{
+public interface UserRepository extends JpaRepository<UserEntity, Integer>{
 
 }
